% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manips.R
\name{split_file_into}
\alias{split_file_into}
\title{Split a file into multiple sheets/files}
\usage{
split_file_into(df, out_file = NULL, split_by = NULL, into = "sheets")
}
\arguments{
\item{out_file}{output file path(s)}

\item{into}{one of "sheets" or "files"}

\item{file}{input file path or a dataframe}
}
\value{
A File is written on provided path(s)
}
\description{
With a file or dataframe as input, possibility to ssplit it on a given variable and write to disk
}
\examples{

split_file_into(iris, out_file = "iris_splitted.xlsx", split_by = "Species", into = "sheets")
split_file_into(iris, out_file = "iris_splitted_\%s.xlsx", split_by = "Species", into = "files")

# Giving a file path works too
rio::export(iris, "iris.xlsx")
split_file_into("iris.xlsx", out_file = "iris_splitted_\%s.xlsx", split_by = "Species", into = "files")

# takes a single file and splits it by the column split_by.
# then write multiple files or a single file with multiple sheets

}
